refactor: block 컴포넌트의 handleInput 파일로 분리
